+++ kinto/plugins/admin/release_hook.py
@@ -6,6 +6,8 @@ in order to build the Kinto Admin UI bundle just before it is packaged.
 * `Documentation <http://zestreleaser.readthedocs.io/en/latest/entrypoints.html>`_
 """
 
+import json
+import os
 import subprocess
 
 
@@ -19,4 +21,12 @@ def after_checkout(data):
 
         The ``node_modules`` folder is excluded using :file:`MANIFEST.in`.
     """
+
+    package_json_path = os.path.join("kinto", "plugins", "admin", "package.json")
+    with open(package_json_path) as f:
+        package_json = json.load(f)
+
+    if package_json["version"] != package_json["dependencies"]["kinto-admin"]:
+        raise ValueError("kinto-admin version mismatch in `package.json`")
+
     subprocess.run(["make", "build-kinto-admin"])